+++ Source/GraphicPipelineCPU/GraphicPipelineCPU.cs
@@ -708,7 +708,7 @@ namespace GraphicPipelineCPU
             return new DrawingHelper2D(prop).GetStringSize(size, text);
         }
 
-        public void DrawString(int x, int y, Color color, float size, string text)
+        public void DrawString(float x, float y, Color color, float size, string text)
         {
             new DrawingHelper2D(prop).DrawString(x, y, color, size, text);
         }
